+++ backend/backend_test.go
@@ -82,15 +82,14 @@ func TestBackendHTTPListener(t *testing.T) {
 			}
 
 			b, err := Initialize(&Config{
-				AgentHost:     "127.0.0.1",
-				AgentPort:     agentPort,
-				APIHost:       "127.0.0.1",
-				APIPort:       apiPort,
-				DashboardHost: "127.0.0.1",
-				DashboardPort: dashboardPort,
-				StateDir:      dataPath,
-				CacheDir:      cachePath,
-				TLS:           tc.tls,
+				AgentHost:        "127.0.0.1",
+				AgentPort:        agentPort,
+				APIListenAddress: fmt.Sprintf("127.0.0.1:%d", apiPort),
+				DashboardHost:    "127.0.0.1",
+				DashboardPort:    dashboardPort,
+				StateDir:         dataPath,
+				CacheDir:         cachePath,
+				TLS:              tc.tls,
 				EtcdAdvertiseClientURLs:      []string{clURL},
 				EtcdListenClientURLs:         []string{clURL},
 				EtcdListenPeerURLs:           []string{apURL},