+++ backend/dashboardd/dashboardd.go
@@ -25,8 +25,7 @@ type Config struct {
 	Port int
 	TLS  *types.TLSOptions
 
-	APIHost string
-	APIPort int
+	APIURL string
 }
 
 // Dashboardd represents the dashboard daemon
@@ -127,7 +126,7 @@ func (d *Dashboardd) Name() string {
 func httpRouter(d *Dashboardd) *mux.Router {
 	r := mux.NewRouter()
 
-	backendProxy, err := newBackendProxy(d.Config.APIHost, d.Config.APIPort, d.Config.TLS)
+	backendProxy, err := newBackendProxy(d.Config.APIURL, d.Config.TLS)
 	if err != nil {
 		d.errChan <- err
 	}
@@ -198,11 +197,11 @@ func noCacheHandler(next http.Handler) http.Handler {
 	})
 }
 
-func newBackendProxy(host string, port int, TLS *types.TLSOptions) (*httputil.ReverseProxy, error) {
+func newBackendProxy(APIURL string, TLS *types.TLSOptions) (*httputil.ReverseProxy, error) {
 	// API gateway to Sensu API
-	target := &url.URL{
-		Host:   fmt.Sprintf("%s:%d", host, port),
-		Scheme: "http",
+	target, err := url.Parse(APIURL)
+	if err != nil {
+		return nil, err
 	}
 
 	// Copy of values from http.DefaultTransport
@@ -221,13 +220,12 @@ func newBackendProxy(host string, port int, TLS *types.TLSOptions) (*httputil.Re
 
 	// Configure TLS
 	if TLS != nil {
-		target.Scheme = "https"
-
 		cfg, err := TLS.ToTLSConfig()
 		if err != nil {
 			return nil, err
 		}
-		cfg.InsecureSkipVerify = true // skip host verification on loopback interface
+		// TODO(palourde): We should avoid using the loopback interface
+		cfg.InsecureSkipVerify = true
 		transport.TLSClientConfig = cfg
 	}