+++ testing/e2e/doc.go
@@ -1,2 +1,0 @@
-// Package e2e contains the end-to-end tests.
-package e2e