+++ util/environment/environment_test.go
@@ -1,16 +1,26 @@
 package environment
 
 import (
+	"fmt"
+	"os"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 )
+
+// Makes platform compliant list of values
+func mkList(key string, s ...string) string {
+	val := strings.Join(s, string(os.PathListSeparator))
+	return fmt.Sprintf("%s=%s", key, val)
+}
 
 func TestMergeEnvironments(t *testing.T) {
 	cases := []struct {
 		name     string
 		env1     []string
 		env2     []string
+		env3     []string
 		expected []string
 	}{
 		{
@@ -45,27 +55,27 @@ func TestMergeEnvironments(t *testing.T) {
 		},
 		{
 			name:     "PATH merge",
-			env1:     []string{"PATH=c:d"},
-			env2:     []string{"PATH=a:b"},
-			expected: []string{"PATH=a:b:c:d"},
+			env1:     []string{mkList("PATH", "c", "d")},
+			env2:     []string{mkList("PATH", "a", "b")},
+			expected: []string{mkList("PATH", "a", "b", "c", "d")},
 		},
 		{
 			name:     "CPATH merge",
-			env1:     []string{"CPATH=c:d"},
-			env2:     []string{"CPATH=a:b"},
-			expected: []string{"CPATH=a:b:c:d"},
+			env1:     []string{mkList("CPATH", "c", "d")},
+			env2:     []string{mkList("CPATH", "a", "b")},
+			expected: []string{mkList("CPATH", "a", "b", "c", "d")},
 		},
 		{
 			name:     "LD_LIBRARY_PATH merge",
-			env1:     []string{"LD_LIBRARY_PATH=c:d"},
-			env2:     []string{"LD_LIBRARY_PATH=a:b"},
-			expected: []string{"LD_LIBRARY_PATH=a:b:c:d"},
+			env1:     []string{mkList("LD_LIBRARY_PATH", "c", "d")},
+			env2:     []string{mkList("LD_LIBRARY_PATH", "a", "b")},
+			expected: []string{mkList("LD_LIBRARY_PATH", "a", "b", "c", "d")},
 		},
 		{
 			name:     "complex example",
-			env1:     []string{"VAR1=VALUE1", "PATH=/bin:/sbin"},
-			env2:     []string{"PATH=~/bin:~/.local/bin", "VAR2=VALUE2"},
-			expected: []string{"VAR1=VALUE1", "VAR2=VALUE2", "PATH=~/bin:~/.local/bin:/bin:/sbin"},
+			env1:     []string{"VAR1=VALUE1", mkList("PATH", "/bin", "/sbin")},
+			env2:     []string{mkList("PATH", "~/bin", "~/.local/bin"), "VAR2=VALUE2"},
+			expected: []string{"VAR1=VALUE1", "VAR2=VALUE2", mkList("PATH", "~/bin", "~/.local/bin", "/bin", "/sbin")},
 		},
 		{
 			name:     "discard invalid environment variables",
@@ -73,11 +83,18 @@ func TestMergeEnvironments(t *testing.T) {
 			env2:     []string{"VAR3="},
 			expected: []string{"VAR2=VALUE2", "VAR3="},
 		},
+		{
+			name:     "more than two sets of variables",
+			env1:     []string{mkList("CPATH", "e", "f"), "VAR1=two"},
+			env2:     []string{mkList("CPATH", "c", "d"), "VAR1=one"},
+			env3:     []string{mkList("CPATH", "a", "b")},
+			expected: []string{mkList("CPATH", "a", "b", "c", "d", "e", "f"), "VAR1=one"},
+		},
 	}
 
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
-			result := MergeEnvironments(tt.env1, tt.env2)
+			result := MergeEnvironments(tt.env1, tt.env2, tt.env3)
 			assert.ElementsMatch(t, result, tt.expected)
 		})
 	}